requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bert-score==0.3.13
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
filelock==3.18.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kagglehub==0.3.10
kiwisolver==1.4.8
litellm==1.63.12
MarkupSafe==3.0.2
matplotlib>=3.5.0
mpmath==1.3.0
multidict==6.2.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
openai==1.67.0
packaging==24.2
pandas>=1.5.0
pillow==11.1.0
polars==1.25.2
portalocker==2.10.1
propcache==0.3.0
protobuf==5.29.4
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
qdrant-client==1.13.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rouge_score==0.1.2
rpds-py==0.23.1
safetensors==0.5.3
six==1.17.0
sniffio==1.3.1
sympy==1.13.1
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
yarl==1.18.3
zipp==3.21.0
seaborn>=0.12.0
